Moved devicemanager out of test file and into source code.
Not complete, but can be integrated.
